reorganized folders into screens and components, created navigation from welcome card to input bee
