+++ App.js
@@ -2,20 +2,20 @@ import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, View } from "react-native";
 import { Provider as PaperProvider } from "react-native-paper";
 import { AppRegistry } from "react-native";
-import NextRound from "./components/NextRound";
-import YesNoCard from "./components/yesNoCard";
-import ChosenCard from "./components/chosenCard";
-import Settings from "./components/settings";
-import Options from "./components/options";
-import WelcomeCard from "./components/welcomeCard";
-import StepOne from "./components/stepOne";
-import StepTwo from "./components/stepTwo";
-import StepThree from "./components/stepThree";
-import NameBee from "./components/nameBee";
-import HowManyCards from "./components/howManyCards";
+import NextRound from "./screens/main/NextRound";
+import YesNoCard from "./screens/main/yesNoCard";
+import ChosenCard from "./screens/main/chosenCard";
+import Settings from "./screens/main/settings";
+import Options from "./screens/main/options";
+import WelcomeCard from "./screens/onBoarding/welcomeCard";
+import StepOne from "./screens/onBoarding/stepOne";
+import StepTwo from "./screens/onBoarding/stepTwo";
+import StepThree from "./screens/onBoarding/stepThree";
+import NameBee from "./screens/onBoarding/nameBee";
+import HowManyCards from "./screens/onBoarding/howManyCards";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import Home from "./components/home";
+import Home from "./screens/main/home";
 import { useFonts } from "expo-font";
 import React from "react";
 import { enableScreens } from "react-native-screens";
@@ -36,13 +36,13 @@ export default function App() {
     <PaperProvider>
       <NavigationContainer>
         <Stack.Navigator>
+          <Stack.Screen name="WelcomeCard" component={WelcomeCard} />
           <Stack.Screen name="Home" component={Home} />
           <Stack.Screen name="YesNoCard" component={YesNoCard} />
           <Stack.Screen name="ChosenCard" component={ChosenCard} />
           <Stack.Screen name="Settings" component={Settings} />
           <Stack.Screen name="NextRound" component={NextRound} />
           <Stack.Screen name="Options" component={Options} />
-          <Stack.Screen name="WelcomeCard" component={WelcomeCard} />
           <Stack.Screen name="StepOne" component={StepOne} />
           <Stack.Screen name="StepTwo" component={StepTwo} />
           <Stack.Screen name="StepThree" component={StepThree} />
