+++ screens/main/chosenCard.js
