+++ screens/onBoarding/stepOne.js
