+++ screens/onBoarding/stepTwo.js
